+++ src/egon/data/airflow/dags/pipeline.py
@@ -9,10 +9,15 @@ from egon.data.datasets import database
 from egon.data.datasets.chp import Chp
 from egon.data.datasets.data_bundle import DataBundle
 from egon.data.datasets.demandregio import DemandRegio
+from egon.data.datasets.district_heating_areas import DistrictHeatingAreas
 from egon.data.datasets.electricity_demand import (
-    CtsElectricityDemand, HouseholdElectricityDemand)
+    CtsElectricityDemand,
+    HouseholdElectricityDemand,
+)
+from egon.data.datasets.electricity_demand_etrago import ElectricalLoadEtrago
 from egon.data.datasets.era5 import WeatherData
 from egon.data.datasets.etrago_setup import EtragoSetup
+from egon.data.datasets.heat_demand import HeatDemandImport
 from egon.data.datasets.heat_etrago import HeatEtrago
 from egon.data.datasets.heat_supply import HeatSupply
 from egon.data.datasets.industry import IndustrialDemandCurves
@@ -31,12 +36,11 @@ from egon.data.datasets.vg250 import Vg250
 from egon.data.datasets.vg250_mv_grid_districts import Vg250MvGridDistricts
 from egon.data.datasets.zensus_mv_grid_districts import ZensusMvGridDistricts
 from egon.data.datasets.zensus_vg250 import ZensusVg250
+from egon.data.datasets.gas_prod import GasProduction
 import airflow
 
-import egon.data.importing.heat_demand_data as import_hd
 import egon.data.importing.zensus as import_zs
 import egon.data.importing.gas_grid as gas_grid
-import egon.data.processing.district_heating_areas as district_heating_areas
 import egon.data.processing.power2gas as power2gas
 import egon.data.processing.substation as substation
 import egon.data.processing.gas_areas as gas_areas
@@ -121,13 +125,9 @@ with airflow.DAG(
     zensus_vg250 = ZensusVg250(
         dependencies=[vg250, population_import])
 
-    zensus_inside_ger_metadata = tasks[
-        'zensus_vg250.add-metadata-zensus-inside-ger']
-
-
     # DemandRegio data import
     demandregio = DemandRegio(dependencies=[
-        setup, vg250, scenario_input_import, data_bundle])
+        setup, vg250, scenario_parameters, data_bundle])
     demandregio_demand_cts_ind = tasks['demandregio.insert-cts-ind-demands']
 
 
@@ -138,19 +138,24 @@ with airflow.DAG(
         population_import,
         zensus_misc_import])
 
-    population_prognosis = tasks['society_prognosis.zensus-population']
-
     # Distribute household electrical demands to zensus cells
     household_electricity_demand_annual = HouseholdElectricityDemand(
-        dependencies=[demandregio, zensus_vg250, zensus_tables,
-                      society_prognosis])
+        dependencies=[
+            demandregio,
+            zensus_vg250,
+            zensus_tables,
+            society_prognosis,
+        ]
+    )
 
     elec_household_demands_zensus = tasks[
-        'electricity_demand.distribute-household-demands']
+        "electricity_demand.distribute-household-demands"
+    ]
 
     # NEP data import
     scenario_capacities = ScenarioCapacities(
-        dependencies=[setup, vg250, data_bundle])
+        dependencies=[setup, vg250, data_bundle]
+    )
     nep_insert_data = tasks["scenario_capacities.insert-data-nep"]
 
     population_import >> nep_insert_data
@@ -215,28 +220,32 @@ with airflow.DAG(
     osmtgmod_substation >> create_voronoi_substation
 
     # MV grid districts
-    mv_grid_districts = mv_grid_districts_setup(dependencies=[create_voronoi_substation])
+    mv_grid_districts = mv_grid_districts_setup(
+        dependencies=[create_voronoi_substation]
+    )
     mv_grid_districts.insert_into(pipeline)
-    define_mv_grid_districts = tasks["mv_grid_districts.define-mv-grid-districts"]
+    define_mv_grid_districts = tasks[
+        "mv_grid_districts.define-mv-grid-districts"
+    ]
 
     # Import potential areas for wind onshore and ground-mounted PV
     re_potential_areas = re_potential_area_setup(dependencies=[setup])
     re_potential_areas.insert_into(pipeline)
-    insert_re_potential_areas = tasks["re_potential_areas.insert-data"]
 
     # Future heat demand calculation based on Peta5_0_1 data
-    heat_demand_import = PythonOperator(
-        task_id="import-heat-demand",
-        python_callable=import_hd.future_heat_demand_data_import,
-    )
-    vg250_clean_and_prepare >> heat_demand_import
-    zensus_inside_ger_metadata >> heat_demand_import
-    scenario_input_import >> heat_demand_import
+    heat_demand_Germany = HeatDemandImport(
+        dependencies=[vg250, scenario_parameters, zensus_vg250])
 
-     # Distribute electrical CTS demands to zensus grid
+    # Distribute electrical CTS demands to zensus grid
     cts_electricity_demand_annual = CtsElectricityDemand(
-        dependencies=[demandregio, zensus_vg250, heat_demand_import,
-                      setup_etrago, household_electricity_demand_annual])
+        dependencies=[
+            demandregio,
+            zensus_vg250,
+            heat_demand_Germany,
+            etrago_input_data,
+            household_electricity_demand_annual,
+        ]
+    )
 
     elec_cts_demands_zensus = tasks[
         'electricity_demand.distribute-cts-demands']
@@ -251,14 +260,13 @@ with airflow.DAG(
     download_data_bundle >> gas_grid_insert_data
     osmtgmod_pypsa >> gas_grid_insert_data
 
-     # Power-to-gas installations creation
+    # Power-to-gas installations creation
     insert_power2gas_installations = PythonOperator(
         task_id="insert-power-to-gas-installations",
         python_callable=power2gas.insert_power2gas,
     )
 
     gas_grid_insert_data >> insert_power2gas_installations
-    osmtgmod_pypsa >> insert_power2gas_installations
 
     # Create gas voronoi
     create_gas_polygons = PythonOperator(
@@ -268,6 +276,10 @@ with airflow.DAG(
 
     gas_grid_insert_data  >> create_gas_polygons
     vg250_clean_and_prepare >> create_gas_polygons
+    
+    # Gas prod import
+    gas_production_insert_data = GasProduction(
+        dependencies=[create_gas_polygons])
 
     # Extract landuse areas from osm data set
     create_landuse_table = PythonOperator(
@@ -287,7 +299,7 @@ with airflow.DAG(
     vg250_clean_and_prepare >> landuse_extraction
 
     # Import weather data
-    weather_data = WeatherData(dependencies=[setup, scenario_parameters])
+    weather_data = WeatherData(dependencies=[setup, scenario_parameters, vg250])
     download_weather_data = tasks["era5.download-era5"]
 
     renewable_feedin = RenewableFeedin(dependencies=[weather_data, vg250])
@@ -297,20 +309,12 @@ with airflow.DAG(
     feedin_solar_thermal = tasks["renewable_feedin.solar-thermal"]
 
     # District heating areas demarcation
-    create_district_heating_areas_table = PythonOperator(
-        task_id="create-district-heating-areas-table",
-        python_callable=district_heating_areas.create_tables
-    )
-    import_district_heating_areas = PythonOperator(
-        task_id="import-district-heating-areas",
-        python_callable=district_heating_areas.
-        district_heating_areas_demarcation
-    )
-    setup >> create_district_heating_areas_table
-    create_district_heating_areas_table >> import_district_heating_areas
+    district_heating_areas = DistrictHeatingAreas(
+        dependencies=[heat_demand_Germany, scenario_parameters])
+    import_district_heating_areas = tasks[
+        "district_heating_areas.demarcation"]
+
     zensus_misc_import >> import_district_heating_areas
-    heat_demand_import >> import_district_heating_areas
-    scenario_input_import >> import_district_heating_areas
 
     # Calculate dynamic line rating for HV trans lines
     calculate_dlr = PythonOperator(
@@ -323,30 +327,43 @@ with airflow.DAG(
 
     # Map zensus grid districts
     zensus_mv_grid_districts = ZensusMvGridDistricts(
-        dependencies=[population_import, mv_grid_districts])
+        dependencies=[population_import, mv_grid_districts]
+    )
 
-    map_zensus_grid_districts = tasks['zensus_mv_grid_districts.mapping']
+    map_zensus_grid_districts = tasks["zensus_mv_grid_districts.mapping"]
 
     # Map federal states to mv_grid_districts
     vg250_mv_grid_districts = Vg250MvGridDistricts(
-        dependencies=[vg250, mv_grid_districts])
+        dependencies=[vg250, mv_grid_districts]
+    )
 
     # Power plants
-    power_plants = PowerPlants(dependencies=[
-        setup, renewable_feedin, mv_grid_districts, mastr_data,
-        re_potential_areas, scenario_parameters, scenario_capacities,
-        Vg250MvGridDistricts, setup_etrago])
+    power_plants = PowerPlants(
+        dependencies=[
+            setup,
+            renewable_feedin,
+            mv_grid_districts,
+            mastr_data,
+            re_potential_areas,
+            scenario_parameters,
+            scenario_capacities,
+            Vg250MvGridDistricts,
+        ]
+    )
 
     power_plant_import = tasks["power_plants.insert-hydro-biomass"]
     generate_wind_farms = tasks["power_plants.wind_farms.insert"]
     generate_pv_ground_mounted = tasks["power_plants.pv_ground_mounted.insert"]
-    solar_rooftop_etrago = tasks["power_plants.pv_rooftop.pv-rooftop-per-mv-grid"]
+    solar_rooftop_etrago = tasks[
+        "power_plants.pv_rooftop.pv-rooftop-per-mv-grid"
+    ]
 
     hvmv_substation_extraction >> generate_wind_farms
     hvmv_substation_extraction >> generate_pv_ground_mounted
     feedin_pv >> solar_rooftop_etrago
     elec_cts_demands_zensus >> solar_rooftop_etrago
     elec_household_demands_zensus >> solar_rooftop_etrago
+    etrago_input_data >> solar_rooftop_etrago
     map_zensus_grid_districts >> solar_rooftop_etrago
 
     # CHP locations
@@ -363,18 +380,9 @@ with airflow.DAG(
 
     # Heat supply
     heat_supply = HeatSupply(
-        dependencies=[data_bundle, chp, zensus_mv_grid_districts])
-
-    import_district_heating_supply = tasks["heat_supply.district-heating"]
-    import_individual_heating_supply = tasks["heat_supply.individual-heating"]
-    heat_supply_tables = tasks["heat_supply.create-tables"]
-    geothermal_potential = tasks["heat_supply.geothermal.potential-germany"]
-
-    create_district_heating_areas_table >> heat_supply_tables
-    import_district_heating_areas >> import_district_heating_supply
-    import_district_heating_areas >> geothermal_potential
-    import_district_heating_areas >> import_individual_heating_supply
-    power_plant_import >> import_individual_heating_supply
+        dependencies=[data_bundle, zensus_mv_grid_districts,
+                      district_heating_areas, power_plants,
+                      zensus_mv_grid_districts, chp])
 
     # Heat to eTraGo
     heat_etrago = HeatEtrago(
@@ -390,5 +398,17 @@ with airflow.DAG(
     )
 
     demand_curves_industry = IndustrialDemandCurves(
-        dependencies=[define_mv_grid_districts, industrial_sites, demandregio_demand_cts_ind, osm, landuse_extraction]
+        dependencies=[
+            define_mv_grid_districts,
+            industrial_sites,
+            demandregio_demand_cts_ind,
+            osm,
+            landuse_extraction,
+        ]
+    )
+
+    # Electrical loads to eTraGo
+
+    electrical_load_etrago = ElectricalLoadEtrago(
+        dependencies=[demand_curves_industry, cts_electricity_demand_annual]
     )